requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
python-logging-logger>=0.5.1.2